Add downgrade catalog option
Puppet 3.x current and future parser produce catalogs that are different
in ways that aren't actually important to the content of the catalog.
The primary differences are the current parser treats all integers as
strings, and never creates a single element array.

Ideally to compare a current and future parser catalog we would upgrade
the current parser catalog, but there just isn't enough information to
be able to do that.

This patch adds a new --downgrade_catalog command-line option that will
preprocess all catalogs given before diffing them to remove spurious
changes in the catalog caused by future parser differences.