Don't search nodes with REST when filter_local param is provided
If one provides the `--filter_local` param it first looked up
with REST and then in local yaml files overriding the earlier results.
Looks like the conditon was accidentially added below and not as elsif.

This changes the condition to only lookup in local cache when
filter_local is provided and default to REST lookup.