Enhance differ for Puppet 4
Without this commit, the "differ" was not using functional routines that
would correctly consume catalogs from disk. Methods like `from_pson` are
no longer available. The "differ" also has implicit code for reading
'.json' extensions despite not having it documented that they were
valid.

This commit removes the ability to attempt to handle JSON catalogs. When
it comes to PSON catalogs, it will use the appropriate and verified
working method available in Puppet 4.

This commit also sorts some of the results of the "diffing" process for
cleaner presentation later.