Clean up "diff" face, present cleaner data
Without this commit, the "diff" face had a few different problems. The
simple ones were a matter of unreadable code. The more important issue
was the lazy method by which data coming out of the "differ" was
formatted and presented.

This commit starts by re-tabbing the whole document. The lines trying to
output to the console were also difficult to read and a bit of a mess.
They were split out into a more usable form.

This commit also takes advantage of the new method of presenting
parameter differences that has been added to the formatter.