Run the param diff formatting correctly
The code that sends the differ output off to be formatted takes a
somewhat simple approach of treating data generically. To present the
param diff data in a more usable manner, a purpose specific method has
been created. We need to make sure this method gets used instead of the
generic formatting approach for this data. Right now the approach taken
for keeping that data from going through the generic method is wrong. A
lack of understanding since the code is sort of dense.

This commit ensures that the param data not be passed through the
generic formatting loop. It also ensures that the param data that gets
passed is actually coming from the correct place.