Load PSON properly between versions of Puppet
This commit adds back the code for loading PSON files and JSON files in
Puppet 3.x. It also ensures that the correct loading method is used
between 3.x and 4.x versions of Puppet. In Puppet 4.x it raises an error
when trying to load JSON catalogs. There was not an obvious method for
loading these files that could be found, additionally a JSON catalog
could not be found to test with.